+++ packages/permissionless/actions/smartAccount/sendTransaction.ts
@@ -113,7 +113,7 @@ export async function sendTransaction<
 
     const callData = await account.encodeCallData({
         to,
-        value: value || 0n,
+        value: value || BigInt(0),
         data: data || "0x"
     })
 
@@ -123,8 +123,8 @@ export async function sendTransaction<
     )({
         userOperation: {
             sender: account.address,
-            maxFeePerGas: maxFeePerGas || 0n,
-            maxPriorityFeePerGas: maxPriorityFeePerGas || 0n,
+            maxFeePerGas: maxFeePerGas || BigInt(0),
+            maxPriorityFeePerGas: maxPriorityFeePerGas || BigInt(0),
             callData: callData,
             nonce: nonce ? BigInt(nonce) : undefined
         },